# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/locations": {
      "get": {
        "operationId": "LocationsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Подстрока по city/state/country",
            "schema": {
              "example": "dush",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Список локаций (cursor, поиск по q)",
        "tags": [
          "locations"
        ]
      },
      "post": {
        "operationId": "LocationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать локацию",
        "tags": [
          "locations"
        ]
      }
    },
    "/api/locations/{id}": {
      "get": {
        "operationId": "LocationsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Одна локация",
        "tags": [
          "locations"
        ]
      },
      "put": {
        "description": "В softclub здесь был 400 (AutoMapper, баг #19). У нас просто работает → 200.",
        "operationId": "LocationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить локацию (полная замена)",
        "tags": [
          "locations"
        ]
      },
      "delete": {
        "operationId": "LocationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить локацию (у постов locationId → null)",
        "tags": [
          "locations"
        ]
      }
    },
    "/api/locations/{id}/posts": {
      "get": {
        "description": "Правила выдачи те же, что у /search/hashtag/{name} (общий PostsService.explore): посты закрытых аккаунтов видны только подписчикам, заблокированные исключены, свои посты в выдачу не попадают. Курсор — id последнего элемента.",
        "operationId": "LocationsController_posts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Лента постов, снятых в этой локации",
        "tags": [
          "locations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Instagram Backend API",
    "description": "NestJS + Prisma + PostgreSQL. Конверт ответа: { data, errors, statusCode }",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UserBriefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userName": {
            "type": "string",
            "example": "eraj"
          },
          "fullName": {
            "type": "string",
            "example": "Eraj Karimov"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          },
          "isPrivate": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "userName",
          "fullName",
          "isVerified",
          "isPrivate"
        ]
      },
...
      "PostMediaDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "http://localhost:9000/instagram/images/2026/07/a.webp"
          },
          "type": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO"
            ],
            "example": "IMAGE"
          },
          "order": {
            "type": "number",
            "example": 0
          },
          "width": {
            "type": "object",
            "nullable": true
          },
          "height": {
            "type": "object",
            "nullable": true
          },
          "duration": {
            "type": "object",
            "nullable": true,
            "description": "Секунды — для видео"
          },
          "thumbUrl": {
            "type": "object",
            "nullable": true,
            "description": "Постер видео"
          },
          "filter": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "url",
          "type",
          "order"
        ]
      },
      "PostLocationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "city": {
            "type": "string",
            "example": "Dushanbe"
          },
          "country": {
            "type": "string",
            "example": "Tajikistan"
          }
        },
        "required": [
          "id",
          "city",
          "country"
        ]
      },
      "PostMusicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 35
          },
          "title": {
            "type": "string",
            "example": "Soundhelix song 1"
          },
          "artist": {
            "type": "string",
            "example": "SoundHelix"
          },
          "streamUrl": {
            "type": "string",
            "example": "http://localhost:3000/api/music/35/stream"
          },
          "coverUrl": {
            "type": "string",
            "example": "http://localhost:9000/instagram/music/covers/x.webp"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "streamUrl",
          "coverUrl"
        ]
      },
      "PostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "caption": {
            "type": "object",
            "nullable": true,
            "maxLength": 2200
          },
          "isReel": {
            "type": "boolean",
            "example": false
          },
          "isArchived": {
            "type": "boolean",
            "example": false
          },
          "author": {
            "description": "Автор — НИКОГДА не null",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserBriefDto"
              }
            ]
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostMediaDto"
            }
          },
          "location": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PostLocationDto"
              }
            ]
          },
          "music": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PostMusicDto"
              }
            ]
          },
          "taggedUsers": {
            "description": "Отмеченные на фото",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserBriefDto"
            }
          },
          "hashtags": {
            "example": [
              "travel",
              "sunset"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "likesCount": {
            "type": "number",
            "example": 24
          },
          "commentsCount": {
            "type": "number",
            "example": 3
          },
          "viewsCount": {
            "type": "number",
            "example": 120
          },
          "isLiked": {
            "type": "boolean",
            "example": false,
            "description": "Лайкнул ли Я"
          },
          "isFavorited": {
            "type": "boolean",
            "example": false,
            "description": "В избранном ли у меня"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isReel",
          "isArchived",
          "author",
          "media",
          "taggedUsers",
          "hashtags",
          "likesCount",
          "commentsCount",
          "viewsCount",
          "isLiked",
          "isFavorited",
          "createdAt"
        ]
      },
      "DeletedDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "LocationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 7
          },
          "city": {
            "type": "string",
            "example": "Dushanbe"
          },
          "state": {
            "type": "object",
            "nullable": true,
            "example": "Sughd"
          },
          "zipCode": {
            "type": "object",
            "nullable": true,
            "example": "734000"
          },
          "country": {
            "type": "string",
            "example": "Tajikistan"
          },
          "lat": {
            "type": "object",
            "nullable": true,
            "example": 38.5598
          },
          "lng": {
            "type": "object",
            "nullable": true,
            "example": 68.787
          }
        },
        "required": [
          "id",
          "city",
          "country"
        ]
      },
      "CreateLocationDto": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "example": "Dushanbe"
          },
          "state": {
            "type": "string",
            "example": "Sughd"
          },
          "zipCode": {
            "type": "string",
            "example": "734000"
          },
          "country": {
            "type": "string",
            "example": "Tajikistan"
          },
          "lat": {
            "type": "number",
            "example": 38.5598
          },
          "lng": {
            "type": "number",
            "example": 68.787
          }
        },
        "required": [
          "city",
          "country"
        ]
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "example": "Dushanbe"
          },
          "state": {
            "type": "string",
            "example": "Sughd"
          },
          "zipCode": {
            "type": "string",
            "example": "734000"
          },
          "country": {
            "type": "string",
            "example": "Tajikistan"
          },
          "lat": {
            "type": "number",
            "example": 38.5598
          },
          "lng": {
            "type": "number",
            "example": 68.787
          }
        },
        "required": [
          "city",
          "country"
        ]
      },
...
    }
  }
}