# src/modules/locations/locations.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  Param,
  ParseIntPipe,
  Post,
  Put,
  Query,
} from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiCreatedResponse,
  ApiOkResponse,
  ApiOperation,
  ApiTags,
} from '@nestjs/swagger';
import { CurrentUser } from '../../common/decorators/current-user.decorator';
import { CursorDto, CursorPage } from '../../common/pagination/cursor.dto';
import { PostDto } from '../posts/dto/post.dto';
import {
  CreateLocationDto,
  DeletedDto,
  LocationDto,
  LocationQueryDto,
  UpdateLocationDto,
} from './dto/location.dto';
import { LocationsService } from './locations.service';

@ApiBearerAuth()
@ApiTags('locations')
@Controller('locations')
export class LocationsController {
  constructor(private readonly locations: LocationsService) {}

  @Get()
  @ApiOperation({ summary: 'Список локаций (cursor, поиск по q)' })
  @ApiOkResponse({ type: LocationDto, isArray: true })
  async list(@Query() dto: LocationQueryDto): Promise<CursorPage<LocationDto>> {
    return this.locations.list(dto);
  }

  @Get(':id')
  @ApiOperation({ summary: 'Одна локация' })
  @ApiOkResponse({ type: LocationDto })
  async get(@Param('id', ParseIntPipe) id: number): Promise<LocationDto> {
    return this.locations.get(id);
  }

  @Get(':id/posts')
  @ApiOperation({
    summary: 'Лента постов, снятых в этой локации',
    description:
      'Правила выдачи те же, что у /search/hashtag/{name} (общий PostsService.explore): ' +
      'посты закрытых аккаунтов видны только подписчикам, заблокированные исключены, ' +
      'свои посты в выдачу не попадают. Курсор — id последнего элемента.',
  })
  @ApiOkResponse({ type: PostDto, isArray: true })
  async posts(
    @CurrentUser('id') userId: string,
    @Param('id', ParseIntPipe) id: number,
    @Query() dto: CursorDto,
  ): Promise<CursorPage<PostDto>> {
    return this.locations.posts(userId, id, dto);
  }

  @Post()
  @ApiOperation({ summary: 'Создать локацию' })
  @ApiCreatedResponse({ type: LocationDto })
  async create(@Body() dto: CreateLocationDto): Promise<LocationDto> {
    return this.locations.create(dto);
  }

  @Put(':id')
  @ApiOperation({
    summary: 'Обновить локацию (полная замена)',
    description: 'В softclub здесь был 400 (AutoMapper, баг #19). У нас просто работает → 200.',
  })
  @ApiOkResponse({ type: LocationDto })
  async update(
    @Param('id', ParseIntPipe) id: number,
    @Body() dto: UpdateLocationDto,
  ): Promise<LocationDto> {
    return this.locations.update(id, dto);
  }

  @Delete(':id')
  @ApiOperation({ summary: 'Удалить локацию (у постов locationId → null)' })
  @ApiOkResponse({ type: DeletedDto })
  async remove(@Param('id', ParseIntPipe) id: number): Promise<DeletedDto> {
    return this.locations.remove(id);
  }
}